homegrid/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.1.1
cycler==0.12.1
distro==1.9.0
docker-pycreds==0.4.0
exceptiongroup==1.2.0
Farama-Notifications==0.0.4
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
gym==0.26.0
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
homegrid==0.1.1
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.4
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.20.3
openai==1.12.0
packaging==21.3
Pillow==9.4.0
protobuf==4.25.1
psutil==5.9.6
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
sacremoses==0.1.1
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.38.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.10.3
torch==2.1.1
torch-ac==1.4.0
torchvision==0.16.1
tqdm==4.66.1
transformers==4.5.1
typing_extensions==4.9.0
urllib3==2.1.0
wandb==0.16.1
zipp==3.17.0
